move firmware finder to new util file. Enable emoji logging for boto3 error.
