+++ urlutil.py
@@ -3,11 +3,6 @@ import logging
 import os.path
 import urllib.request
 from urllib.parse import urlparse
-try:
-    import boto3
-    from botocore.handlers import disable_signing
-except ImportError:
-    logging.warn("commit hash param will not work without boto3")
 
 def get_json_from_url(url):
     with urllib.request.urlopen(url) as stream:
@@ -31,16 +26,6 @@ def get_local_file_from_url(url, tempdir):
             output.write(stream.read())
 
     return pathname
-
-def find_firmware_by_hash(hash, board, language):
-    resource = boto3.resource('s3')
-    resource.meta.client.meta.events.register('choose-signer.s3.*', disable_signing)
-    bucket = resource.Bucket('adafruit-circuit-python')
-
-    for item in bucket.objects.filter(Prefix="bin/{}/{}/".format(board, language)):
-        if hash in item.key and ".uf2" in item.key:
-            return item.key
-    return None
 
 
 def get_s3_url(s3_path):
